Code refactored and background added
