+++ client/src/pages/Home/Landing.jsx
@@ -1,39 +1,23 @@
-import Card from "../../components/Card/Card2";
-import HomeNav from "../../components/Home/HomeNav";
+import Card from "../../components/Home/Card/CardLatest";
+import HomeNav from "../../components/Home/HomeNav/HomeNav";
 import HomeLayout from "../../layout/HomeLayout";
-import PostCard from "./Card";
+import PostCard from "../../components/Home/Card/PostCard";
+import HeroSection from "../../components/Home/HeroSection/HeroSection";
 // import LatestCard from "./LatestCard";
 export default function Landing({ title }) {
     return (
         <HomeLayout>
             <div className="flex flex-col justify-center items-center h-full w-full overflow-hidden bg-black">
                 <div className="w-[95%]">
-                    <div className="w-full h-[70%] bg-grid-pattern bg-[length:139px_152px] bg-[#0B0D18]">
-                        <div className="flex max-w-full">
-                            <div className=" max-h-full py-32 px-20">
-                                <h1 className="text-6xl font-extralight text-white text-wrap m-8">Innovate, Learn, Connect</h1>
-                                <p className="max-w-full text-xl text-white font-normal text-wrap mx-8">Explore the latest tech trends, gain insights from industry experts, and engage with fellow tech enthusiasts. From AI breakthroughs to coding tips, our blog is your go-to resource for all things.</p>
-                                <div className=" flex justify-start items-center w-full">
-                                    <button
-                                        className="bg-blue-500 text-white max-w-full w-28 h-10 rounded-3xl m-8 hover:bg-[#65558F] ease-out">Sign Up</button>
-                                    <button
-                                        className="bg-[#A3C8F0] text-black max-w-full w-28 h-10 rounded-3xl">Write</button>
-                                </div>
-                            </div>
-                            <div
-                                className="flex items-end w-[900px] max-h-full mt-28">
-                                <img src="/Saly-11.png"
-                                    className="h-full " />
-                            </div>
-                        </div>
-                    </div>
+                    <HeroSection/>
                     <HomeNav />
                     <div>
                         <PostCard logo={"/Profilelogo.png"} image={"/CardImg.png"} author={"Matthew"} category="Free Factor" title={"How Did a Generation Become So Stupid "} description={"Lorem ipsum dolor sit amet consectetur adipisicing elit. Harum ratione nesciunt sunt quaerat vero. At nihil ad repellendus quidem voluptas."} date={"Jun 8"} likes={"10.5K"} comments={241} />
                         <PostCard logo={"/Profilelogo.png"} image={"/CardImg.png"} author={"Matthew"} category="Free Factor" title={"How Did a Generation Become So Stupid "} description={"Lorem ipsum dolor sit amet consectetur adipisicing elit. Harum ratione nesciunt sunt quaerat vero. At nihil ad repellendus quidem voluptas."} date={"Jun 8"} likes={"10.5K"} comments={241} />
                         <PostCard logo={"/Profilelogo.png"} image={"/CardImg.png"} author={"Matthew"} category="Free Factor" title={"How Did a Generation Become So Stupid "} description={"Lorem ipsum dolor sit amet consectetur adipisicing elit. Harum ratione nesciunt sunt quaerat vero. At nihil ad repellendus quidem voluptas."} date={"Jun 8"} likes={"10.5K"} comments={241} />
                     </div>
-                    <div className="bg-[#0B0D18] w-full h-full flex flex-wrap items-center justify-evenly py-10  gap-y-6">
+                    <h1 className="text-white text-3xl mb-2">Latest</h1>
+                    <div className="bg-[#0B0D18] w-full h-full flex flex-wrap items-center justify-evenly py-10  gap-y-6 border-t-2 border-red-600 p-4 rounded-md">
                         <Card logo={"/Profilelogo.png"} author={"Matthew"} category="Free Factor" title={"How Did a Generation Become So Stupid? "} date={"Jun 8"} likes={"10.5K"} comments={241} />
                         <Card logo={"/Profilelogo.png"} author={"Matthew"} category="Free Factor" title={"How Did a Generation Become So Stupid? "} date={"Jun 8"} likes={"10.5K"} comments={241} />
                         <Card logo={"/Profilelogo.png"} author={"Matthew"} category="Free Factor" title={"How Did a Generation Become So Stupid? "} date={"Jun 8"} likes={"10.5K"} comments={241} />
